KroneckerProducts

Index

Base.Matrix

KroneckerProducts.BoxProduct

KroneckerProducts.KroneckerProduct

Base.adjoint

Base.adjoint

Base.inv

Base.inv

Base.kron

Base.transpose

Base.transpose

IgaBase.order

KroneckerProducts.box

KroneckerProducts.issquare

LinearAlgebra.det

LinearAlgebra.det

LinearAlgebra.eigen

LinearAlgebra.isposdef

LinearAlgebra.issymmetric

LinearAlgebra.lmul!

LinearAlgebra.logdet

LinearAlgebra.rmul!

LinearAlgebra.tr

Base.adjoint — Method

adjoint(K::BoxProduct)

Compute the adjoint of a Box product.

Base.adjoint — Method

adjoint(K::KroneckerProduct)

Compute the adjoint of a Kronecker product.

Base.inv — Method

inv(K::BoxProduct)

Compute the inverse of a Box product.

Base.inv — Method

inv(K::KroneckerProduct)

Compute the inverse of a Kronecker product.

Base.kron — Method

collect(K::KroneckerProduct)

Collects a lazy instance of the KroneckerProduct type into a full, native matrix. Equivalent with Matrix(K::KroneckerProduct).

Base.transpose — Method

transpose(K::BoxProduct)

Compute the transpose of a Box product.

Base.transpose — Method

transpose(K::KroneckerProduct)

Compute the transpose of a Kronecker product.

IgaBase.order — Method

order(M::AbstractMatrix)

Returns the order of a matrix, i.e. how many matrices are involved in the Kronecker product (default to 1 for general matrices).

KroneckerProducts.box — Function

box(A, B)
box(A, B...)

Compute the box product, which is defined as a perfect shuffle of a kronecker product.

KroneckerProducts.issquare — Method

issquare(A::AbstractMatrix)

Checks if an array is a square matrix.

LinearAlgebra.det — Method

det(K::BoxProduct)

Compute the determinant of a Box product.

LinearAlgebra.det — Method

det(K::KroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.eigen — Method

eigen(K::KroneckerProduct)

If the matrices of a KroneckerProduct instance are square, performs the eigenvalue decompositon on them and return an Eigen type.

LinearAlgebra.isposdef — Method

isposdef(K::KroneckerProduct)

Test whether a Kronecker product is positive definite (and Hermitian) by trying to perform a Cholesky factorization of K.

LinearAlgebra.issymmetric — Method

issymmetric(K::KroneckerProduct)

Checks if a Kronecker product is symmetric.

LinearAlgebra.lmul! — Method

lmul!(a::Number, K::KroneckerProduct)

Scale a KroneckerProduct K inplace by a factor a by rescaling the left matrix.

LinearAlgebra.logdet — Method

logdet(K::KroneckerProduct)

Compute the logarithm of the determinant of a Kronecker product.

LinearAlgebra.rmul! — Method

rmul!(K::KroneckerProduct, a::Number)

Scale a KroneckerProduct K inplace by a factor a by rescaling the right matrix.

LinearAlgebra.tr — Method

tr(K::KroneckerProduct)

Compute the trace of a Kronecker product.

Base.Matrix — Method

Matrix(K::KroneckerProduct)

Converts a GeneralizedKroneckerProduct instance to a Matrix type.

KroneckerProducts.BoxProduct — Type

BoxProduct{M,T,N} <: AbstractKron{M,T,N}

The box product is defined in the following paper.

Olsen, Peder A., Steven J. Rennie, and Vaibhava Goel. "Efficient automatic differentiation of matrix functions." In Recent Advances in Algorithmic Differentiation, pp. 71-81. Springer, Berlin, Heidelberg, 2012.

KroneckerProducts.KroneckerProduct — Type

KroneckerProduct{T,N,S}  <: AbstractVecOrMat{T}

Matrix type that stores the matrices in a Kronecker product and lazily evaluates products, sums, etc.